Prevent duplicate entries from being added to the config
Basically, if someone cloned a repo, deleted the folder contents, and then cloned a repo to the same folder, duplicate repo entries would be created for the same folder. I initially thought this might be a niche edge-case, but then I thought of some scenarios where it might happen (like deleting and re-cloning a repository, because something went wrong).