Add repo to roles constructor | get actual list of roles for repo
I had to move the manage roles button from the header and add it to the repo list. I didn't realize the roles list would need to be repo-sensitive at first, but it needs to be.

Also, add logic to get the real list of roles from the repo. It's great - everything's working now!